// src/common/interfaces/Thema.ts
// Generated by https://quicktype.io

export interface Thema {
  palette: Palette
  transitions: Transitions
}

export interface Palette {
  mode: string
  header: Header
  menuLateral: MenuLateral
  tabs: Tabs
  primary: Color
  secondary: Color
  error: Color
  warning: Color
  info: Color
  success: Color
  divider: string
  action: Action
  grey: Grey
  text: Text
}

export interface Action {
  hover: string
  disabled: string
  selected: string
}

export interface Color {
  main: Colorvalues
  light: Colorvalues
  dark: Colorvalues
}

export interface Colorvalues {
  value: string
  contrastText: string
}

export interface Grey {
  '100': string
}

export interface Header {
  background: string
  logo: string
  texto: Texto
  icones: string
}

export interface Texto {
  color: string
}

export interface MenuLateral {
  background: string
  icones: Icone
}

export interface Icone {
  color: string
  size: string
}

export interface Tabs {
  background: string
  iconeHome: Icone
}

export interface Text {
  primary: string
  secondary: string
}

export interface Transitions {
  buttom: string
}